do not wipe CFLAGS in configure.ac
When extra CFLAGS are passed in the environment for ./configure, these
are expected to be used while compiling. By setting `CFLAGS=` to an
empty value in configure.ac, distribution specific CLFAGS will not be
available anymore.

This caused linking issues on Fedora as it tries to compile with extra
hardening flags.

Signed-off-by: Niels de Vos <ndevos@redhat.com>